% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straw.R
\name{create_parse_graph}
\alias{create_parse_graph}
\title{Produce an ast or parse graph of an expression}
\usage{
create_parse_graph(expr, type = c("parse", "ast")[1])
}
\arguments{
\item{expr}{An expression (e.g. created by parse(text = "1 + 1"))}

\item{type}{Generate a full parse graph or ast}
}
\value{
An igraph graph object
}
\description{
Produce an ast or parse graph of an expression
}
\examples{
create_parse_graph(parse(text = "if (x) 2 else a$b"))
create_parse_graph(expression(if (x) 2 else a$b), type = "ast")

}
