% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straw.r
\name{match_expr}
\alias{match_expr}
\title{Produce an ast or parse graph of an expression}
\usage{
match_expr(g_parse, g_ast, parse_node, ast_node)
}
\arguments{
\item{g_parse}{Parse graph generated by create_parse_graph}

\item{g_ast}{AST graph generated by create_parse_graph}

\item{parse_node}{node from g_parse to search from}

\item{ast_node}{node from g_ast to search from}
}
\value{
A data.frame linking the graphs with columns parse_id, ast_id
}
\description{
Produce an ast or parse graph of an expression
}
\examples{
exp <- parse(text = "if (x) 2 else a$'b'")
g <- create_parse_graph(parse(text = "if (x) 2 else a$'b'"))
plot(g, layout = layout.reingold.tilford)

g_ast <- create_parse_graph(exp, type = "ast")
plot(g_ast, layout = layout.reingold.tilford)

match_expr(g, g_ast, g[2,], g_ast[2,])

}
