% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straw.r
\name{match_path}
\alias{match_path}
\title{Produce an ast or parse graph of an expression}
\usage{
match_path(g_parse, g_ast, row_num)
}
\arguments{
\item{g_parse}{Parse graph generated by create_parse_graph}

\item{g_ast}{AST graph generated by create_parse_graph}

\item{id}{id of ast node to find}
}
\value{
A path of vectors in g_parse leading to corresponding ast node
}
\description{
Produce an ast or parse graph of an expression
}
\examples{
exp <- parse(text = "if (x) 2 else a$'b'")
g <- create_parse_graph(parse(text = "if (x) 2 else a$'b'"))
plot(g, layout = layout.reingold.tilford)

g_ast <- create_parse_graph(exp, type = "ast")
plot(g_ast, layout = layout.reingold.tilford)

# find code corresponding to `$` call
match_path(g, g_ast, V(g_ast)[label == "$"]$id)
match_path(g, g_ast, 8)
}
