% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multimatch}
\alias{multimatch}
\title{Produce an ast or parse graph of an expression}
\usage{
multimatch(x, table)
}
\arguments{
\item{x}{values to be matched}

\item{table}{values to be matched against}
}
\value{
list with vector of matching table indices for each element in x
}
\description{
Produce an ast or parse graph of an expression
}
\examples{
multimatch(1:3, c(1,2,3,1,2,3))

}
