% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straw.R
\name{plot_match}
\alias{plot_match}
\title{Produce an ast or parse graph of an expression}
\usage{
plot_match(g_parse, g_ast, id)
}
\arguments{
\item{g_parse}{Parse graph generated by create_parse_graph}

\item{g_ast}{AST graph generated by create_parse_graph}

\item{id}{id of ast node to find}
}
\value{
A path of vectors in g_parse leading to corresponding ast node
}
\description{
Produce an ast or parse graph of an expression
}
\examples{
exp <- parse(text = "if (x) 2 else a$'b'")
g <- create_parse_graph(parse(text = "if (x) 2 else a$'b'"))
plot(g, layout = layout.reingold.tilford)

g_ast <- create_parse_graph(exp, type = "ast")
plot(g_ast, layout = layout.reingold.tilford)

# find code corresponding to `$` call
plot_match(g, g_ast, 8)

}
