% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{straw_R}
\alias{straw_R}
\title{Straw Quick Dump}
\usage{
straw_R(argv)
}
\arguments{
\item{argv}{string of arguments separated by spaces}
}
\value{
Data.frame of a sparse matrix of data from hic file. x,y,counts
}
\description{
fast C++ implementation of dump. Not as fully featured as the
Java version. Reads the .hic file, finds the appropriate matrix and slice
of data, and outputs as data.frame in sparse upper triangular format.
Currently only supporting matrices.
}
\details{
Usage: straw_R <NONE/VC/VC_SQRT/KR> <hicFile(s)> <chr1>[:x1:x2] <chr2>[:y1:y2] <BP/FRAG> <binsize>
}
