% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strawr.R
\docType{package}
\name{strawr}
\alias{strawr}
\title{strawr}
\description{
API for fast data extraction for .hic files that provides programmatic access to the matrices. It doesn't store the pointer data for all the matrices, only the one queried, and currently we are only supporting matrices (not vectors).
}
